// packages/coin-solana/src/sdk/metaplex/mpl-token-metadata/types/DelegateArgs.ts
export type DelegateArgsRecord = {
  CollectionV1: { authorizationData: beet.COption<AuthorizationData> };
  SaleV1: { amount: beet.bignum; authorizationData: beet.COption<AuthorizationData> };
  TransferV1: { amount: beet.bignum; authorizationData: beet.COption<AuthorizationData> };
  DataV1: { authorizationData: beet.COption<AuthorizationData> };
  UtilityV1: { amount: beet.bignum; authorizationData: beet.COption<AuthorizationData> };
  StakingV1: { amount: beet.bignum; authorizationData: beet.COption<AuthorizationData> };
  StandardV1: { amount: beet.bignum };
  LockedTransferV1: {
    amount: beet.bignum;
    lockedAddress: web3.PublicKey;
    authorizationData: beet.COption<AuthorizationData>;
  };
  ProgrammableConfigV1: { authorizationData: beet.COption<AuthorizationData> };
  AuthorityItemV1: { authorizationData: beet.COption<AuthorizationData> };
  DataItemV1: { authorizationData: beet.COption<AuthorizationData> };
  CollectionItemV1: { authorizationData: beet.COption<AuthorizationData> };
  ProgrammableConfigItemV1: { authorizationData: beet.COption<AuthorizationData> };
};

/**
 * Union type respresenting the DelegateArgs data enum defined in Rust.
 *
 * NOTE: that it includes a `__kind` property which allows to narrow types in
 * switch/if statements.
 * Additionally `isDelegateArgs*` type guards are exposed below to narrow to a specific variant.
 *
 * @category userTypes
 * @category enums
 * @category generated
 */
export type DelegateArgs = beet.DataEnumKeyAsKind<DelegateArgsRecord>;