// packages/coin-solana/src/sdk/metaplex/mpl-token-metadata/types/TokenDelegateRole.ts
/**
 * The following methods are based on `mpl-token-metadata`, thanks for their work
 * https://github.com/metaplex-foundation/mpl-token-metadata/tree/main/programs/token-metadata/js/src/generated/types
 */

/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet';
/**
 * @category enums
 * @category generated
 */
export enum TokenDelegateRole {
  Sale,
  Transfer,
  Utility,
  Staking,
  Standard,
  LockedTransfer,
  Migration,
}

/**
 * @category userTypes
 * @category generated
 */
export const tokenDelegateRoleBeet = beet.fixedScalarEnum(TokenDelegateRole) as beet.FixedSizeBeet<
  TokenDelegateRole,
  TokenDelegateRole
>;